refactor : #10 아키텍처 리팩토링 및  디렉토리 구조 변경
