--- src/main/java/org/farm/fireflyserver/care/domain/Authority.java
@@ -1,5 +1,0 @@
-package org.farm.fireflyserver.care.domain;
-
-public enum Authority {
-    MNG, GOV, DEV
-}
