--- src/main/java/org/farm/fireflyserver/care/domain/CareManager.java
@@ -1,24 +1,0 @@
-package org.farm.fireflyserver.care.domain;
-
-import jakarta.persistence.*;
-import lombok.AccessLevel;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-import org.hibernate.annotations.Comment;
-
-@Entity
-@Getter(AccessLevel.PROTECTED)
-@NoArgsConstructor
-public class CareManager {
-
-    @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
-
-    @OneToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "user_id")
-    private Account account;
-
-    @Comment("전화 번호")
-    private String phoneNum;
-}
