--- src/main/java/org/farm/fireflyserver/care/domain/Result.java
@@ -1,5 +1,0 @@
-package org.farm.fireflyserver.care.domain;
-
-public enum Result {
-    NORMAL, ABSENT
-}
