--- src/main/java/org/farm/fireflyserver/care/domain/Type.java
@@ -1,7 +1,0 @@
-package org.farm.fireflyserver.care.domain;
-
-public enum Type {
-    CALL,
-    VISIT,
-    EMERGENCY
-}
