+++ src/main/java/org/farm/fireflyserver/domain/care/domain/Authority.java
@@ -1,0 +1,5 @@
+package org.farm.fireflyserver.domain.care.domain;
+
+public enum Authority {
+    MNG, GOV, DEV
+}
