+++ src/main/java/org/farm/fireflyserver/domain/care/domain/Result.java
@@ -1,0 +1,5 @@
+package org.farm.fireflyserver.domain.care.domain;
+
+public enum Result {
+    NORMAL, ABSENT
+}
