+++ src/main/java/org/farm/fireflyserver/domain/care/domain/Type.java
@@ -1,0 +1,7 @@
+package org.farm.fireflyserver.domain.care.domain;
+
+public enum Type {
+    CALL,
+    VISIT,
+    EMERGENCY
+}
