+++ src/main/java/org/farm/fireflyserver/domain/led/persistence/entity/SensorGbn.java
@@ -1,0 +1,20 @@
+package org.farm.fireflyserver.domain.led.persistence.entity;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+@RequiredArgsConstructor
+@Getter
+public enum SensorGbn {
+
+    BED_ROOM("01", "안방"),
+    LIVING_ROOM("02", "거실"),
+    KITCHEN("03", "주방"),
+    TOILET("04", "화장실");
+
+    private final String code;
+    private final String desc;
+
+    //코드로 저장시 Convert 필요
+
+}
