+++ src/main/java/org/farm/fireflyserver/domain/senior/persistence/entity/Gender.java
@@ -1,0 +1,5 @@
+package org.farm.fireflyserver.domain.senior.persistence.entity;
+
+public enum Gender {
+    MALE, FEMALE
+}
