+++ src/main/java/org/farm/fireflyserver/domain/senior/persistence/repository/SeniorRepository.java
@@ -1,0 +1,8 @@
+package org.farm.fireflyserver.domain.senior.persistence.repository;
+
+import org.farm.fireflyserver.domain.senior.persistence.entity.Senior;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface SeniorRepository extends JpaRepository<Senior, Long> {
+
+}
