--- src/main/java/org/farm/fireflyserver/senior/application/port/in/RegisterSeniorUseCase.java
@@ -1,7 +1,0 @@
-package org.farm.fireflyserver.senior.application.port.in;
-
-import org.farm.fireflyserver.senior.application.command.RegisterSeniorCommand;
-
-public interface RegisterSeniorUseCase {
-    void registerSenior(RegisterSeniorCommand command);
-}
