--- src/main/java/org/farm/fireflyserver/senior/domain/Gender.java
@@ -1,5 +1,0 @@
-package org.farm.fireflyserver.senior.domain;
-
-public enum Gender {
-    MALE, FEMALE
-}
